+++ src/parser/mod.rs
@@ -306,11 +306,6 @@ impl BgpkitParser<Box<dyn Read + Send>> {
 }
 
 impl<R> BgpkitParser<R> {
-    /// Return the filters currently configured on this parser.
-    pub fn filters(&self) -> &[Filter] {
-        &self.filters
-    }
-
     pub(crate) fn warn_zebra_compat_once(&mut self) {
         self.options.warn_zebra_compat_once();
     }